split test suite into more manageable parts
